Add files via upload
Let me know if this works. I downloaded the script from Jupyter as a .py file then transferred it to a new folder to organize it. I just dragged and dropped it into the repository through the upload feature and I'm hoping you can view and even edit the script