Add files via upload
This is a combination of the z-axis vector generator/classifier and FF MLP scripts we all worked on. This has a general network structure that is very easy to vary the layers paired with a sigmoid activation function. The data can generate n number of 3 dimensional vectors and classify them as being above or below the x-y plane.